// mobile/src/themes/types.d.ts
import { Theme } from "react-native-paper/lib/typescript/src/types";

/* import { ColorValue } from "react-native";

type FontWeights =
	| "normal"
	| "bold"
	| "100"
	| "200"
	| "300"
	| "400"
	| "500"
	| "600"
	| "700"
	| "800"
	| "900";

type FontProps = {
	fontFamily: string;
	fontWeight: FontWeights;
};

type FontsProps = {
	light: FontProps | string;
	medium: FontProps | string;
	regular: FontProps | string;
	thin: FontProps | string;
};

export type Theme = {
	dark: boolean;
	colors: {
		primary: ColorValue;
		secundary: ColorValue;

		textOnPrimary: ColorValue;
		textOnSecundary: ColorValue;

		disabledBackground: ColorValue;
		disabledText: ColorValue;

		placeholder: ColorValue;
	};

	fonts: string | FontsProps;
};
 */

export type Theme = Theme;